backend/requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
pymysql
python-multipart
cryptography
requests
python-dotenv
numpy
openpyxl
chromadb